requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openai>=1.40.0
jieba==0.42.1
openai-whisper==20231117
werkzeug==3.0.1
yt-dlp>=2023.12.30
trafilatura>=1.6.0
google-genai>=0.2.0
llama-index>=0.10.0
llama-index-vector-stores-postgres>=0.1.0
llama-index-embeddings-google-genai>=0.1.0
llama-index-llms-google-genai>=0.1.0
psycopg2-binary>=2.9.0

